Begin to convert to bank switching
Initial conversion to bank switching.  EG0 model is mapped to bank 0, 0x20 (i.e. 0x020), which is the same as the "original" variables.  EG1 model is bank 2, 0x20 (i.e. 0x120).
EG0 still works.  The bank switching is completely untested.  Also need to do work with gates for this to work.