Rework MainLoop
First, rework to put the common code at the top of the loop to:
increment the ADC_CHANNEL
do conversion (into ADC_VALUE)
extract the  FADERACTIVE_FLAG and
MODEL_VALUE for later evaluation.
more yet to do, but this change is done.  Now evaluate those values in the dedicated routines for A/D/S/R